requirements.txt
wikipedia==1.4.0
SQLAlchemy==1.4.32
telebot==0.0.4
requests==2.27.1
imdbpy==0.1.5
kinopoiskpy==0.9.0
kinopoisk-api-unofficial-client==2.1.1
